+++ internal/tui/terminal_default.go
@@ -1,0 +1,9 @@
+//go:build !windows
+
+package tui
+
+import "os"
+
+func setupTerminalOutput(_ *os.File) (func(), bool) {
+	return func() {}, true
+}